# api.yaml
swagger: "2.0"
info:
  description: "Курсовой проект 'Ветеринарная клиника' "
  version: "1.0.0"
  title: "Ветеринарная клиника"
  contact:
    email: "minaevam87@gmail.com"
host: "petstore.swagger.io"
basePath: "/veterinary/v1"
...
schemes:
  - "https"
paths:
  /auth/login:
    post:
      tags:
        - "user"
      summary: "Login"
      description: "This can only be done by the logged in user."
      operationId: "loginUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User credentials"
          required: true
          schema:
            $ref: "#/definitions/UserCredentials"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: '#/definitions/AuthorizedUser'
        "400":
          description: "invalid credentials"
          schema:
            $ref: "#/definitions/Error"
  /auth/register:
    post:
      tags:
        - "user"
      summary: "Registration"
      description: ""
      operationId: "register"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User data"
          required: true
          schema:
            $ref: "#/definitions/UserData"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: '#/definitions/User'
        "400":
          description: "invalid user data"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  UserCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      firstName:
        type: "string"
      lastName:
        type: "string"
      patronymic:
        type: "string"
      fio:
        type: "string"
      userPic:
        type: "string"
      email:
        type: "string"
      status:
        type: "string"
      phone:
        type: "string"
      lastVisit:
        type: "string"
      roles:
        type: "array"
        items:
          $ref: "#/definitions/Role"
  UserData:
    type: "object"
    properties:
      firstName:
        type: "string"
      patronymic:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      userPic:
        type: "string"
      password:
        type: "string"
      locale:
        type: "string"
      gender:
        $ref: "#/definitions/Gender"
  AuthorizedUser:
    type: "object"
    properties:
      userName:
        type: "string"
      token:
        type: "string"
      roles:
        type: "array"
        items:
          $ref: "#/definitions/Role"
  Gender:
    type: "object"
    properties:
      value:
        type: "string"
  Role:
    type: "object"
    properties:
      value:
        type: "string"
...
  Error:
    type: "object"
    properties:
      code:
        type: "string"
      message:
        type: "string"
...